fix rgw_lua.h dependency
