+++ src/rgw/rgw_lua_utils.h
@@ -518,98 +518,4 @@ struct StringMapMetaTable : public EmptyMetaTable {
   }
 };
 
-// Metadata maintained about each lua script
-struct LuaScriptMeta {
-  
-  // The priority of the script. Scripts are executed from lowest priority to highest 
-  int priority;
-
-  // The context for which the script is run
-  context ctx;
-
-  // script name provided by the user
-  std::string name;
-
-  // Actual script contents
-  std::string script;
-
-  LuaScriptMeta(
-    int priority, 
-    context ctx, 
-    std::string name,
-    std::string script
-  ): priority(priority), ctx(ctx), name(name), script(script) {}
-
-  LuaScriptMeta(std::string script, context ctx): priority(MAX_LUA_PRIORITY), ctx(ctx), name(""), script(script) {}
-
-  void print() {
-    std::cout << "Script Name: " << name << std::endl;
-    std::cout << "Priority: " << priority << std::endl;
-    std::cout << "Context: " << to_string(ctx) << std::endl;
-    std::cout << "Script Contents:" << std::endl;
-    std::cout << "==========================================" << std::endl;
-    std::cout << script << std::endl;
-    std::cout << "==========================================" << std::endl << std::endl;
-  }
-
-  void encode(bufferlist& bl) const {
-    ENCODE_START(1, 1, bl);
-    encode(priority, bl);
-    encode(to_string(ctx), bl);
-    encode(name, bl);
-    encode(script, bl);
-    ENCODE_FINISH(bl);
-  }
-
-  void decode(bufferlist::const_iterator& bl) {
-    string stored_ctx;
-    
-    DECODE_START(1, bl);
-    decode(priority, bl);
-    decode(stored_ctx, bl);
-    ctx = to_context(stored_ctx);
-    decode(name, bl);
-    decode(script, bl);
-    DECODE_FINISH(bl);
-  }
-
-}
-
-// Array of LuaScriptMeta for all scripts
-struct LuaRuntimeMeta {
-  std::vector<LuaScriptMeta> scripts;
-
-  LuaRuntimeMeta() {
-    scripts = std::vector<LuaScriptMeta>;
-  }
-
-  LuaRuntimeMeta(LuaScriptMeta script) {
-    scripts = std::vector<LuaScriptMeta>(script);
-  }
-
-  void encode(bufferlist& bl) const {
-    ENCODE_START(1, 1, bl);
-    encode(scripts.size(), bl);
-    for (auto script : scripts) {
-      encode(script);
-    }
-    ENCODE_FINISH(bl);
-  }
-
-  void decode(bufferlist::const_iterator& bl) {
-    int num_scripts;
-    std::string script;
-
-    DECODE_START(1, bl);
-    decode(num_scripts, bl);
-    scripts = std::vector<LuaScriptMeta>(num_scripts);
-
-    for (int idx = 0; idx < num_scripts; idx++) {
-      decode(script, bl);
-      scripts[idx] = script;
-    }
-    DECODE_FINISH(bl);
-  }
-}
-
 } // namespace rgw::lua
